enh: Improve the code generation scripts
Allow to set the project root directory via an environment variable.
This helps people with a slightly different git checkout schema. After
the introduction of Go modules, one is not required anymore to follow the
GOPATH conventions with the /src/ subdirectory.

Moreover, only start netbox to fetch the swagger spec. This helps, to iterate quickly,
if one wants to develop a new patch for the swagger spec.

Last but not least, set some bash debugging options.